+++ frontend/src/helpers/auth.validate.ts
